# dist/api.yml
openapi: 3.0.0
info:
  title: api
  description: API for project
  version: 1.0.0
...
paths:
...
  /admin/addPreviewMatch:
    post:
      tags:
      - Admin
      summary: Add New match to system
      description: The admin add's details about a new future game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewMatch'
        description:
      responses:
        "201":
          description: Add New match to system Successfuly
        "400":
          description: fail to add match to system
  /admin/addMatch:
    post:
      tags:
      - Admin
      summary: Add New match to the system
      description: The admin add's details about a new future game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Match'
        description:
      responses:
        "201":
          description:  a New match is added to the system Successfuly
        "400":
          description: failed to add a match to the system
  /admin/addScoretoMatch:
    put:
      tags:
      - Admin
      summary: Add a Score to a match.
      description: The admin add's a score to a match
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                MatchId:
                  type: int
                  description: an id for a game the admin already added
                  example: 14
                HomeGoals:
                  type: string
                  description: amount of goals for thr hometeam
                  example: 1
                AwayGoals:
                  type: string
                  description: amount of goals for thr awayteam
                  example: 1
      responses:
        '200':
          description: A new score has been added to game
        '400':
          description: Wrong input parameters
  /admin/addEventtoMatch:
    post:
      tags:
      - Admin
      summary: Add New event to a match.
      description: The admin add's events to a match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventDetails'
        description: add a new event
      responses:
        '200':
          description: A new event has been added.
        '400':
          description: failed to add an event.
...
components:
  schemas:
...
    EventInGame:
      type: string
      enum:
      - Goal
      - Offside
      - Foul
      - RedCard
      - YellowCard
      - Injury
      - SubPlayer
    EventDetails:
      type: object
      properties:
        MatchId:
          type: integer
          example: 80
        Date:
          type: string
          format: Date
          example: 2017-07-21T00:00:00.000+0000
        Hour:
          type: string
          example: 15:30
        TimeMinuteInGame:
          maximum: 90
          minimum: 0
          type: integer
          example: 60
        EventInGame:
          type: string
          $ref: '#/components/schemas/EventInGame'
          description: You need to choose one of the following options - Goal, Offside, Foul,RedCard,YellowCard,Injury,SubPlayer
        player_id:
          type: integer
          example: Neymar
        Description:
          type: string
          example: 15.05.2021 21:30 60, Red card to Liad Segev
    PreviewMatch:
      type: object
      properties:
        date:
          type: string
          #format: date
        hour:
          type: string
          example: "15:30"
        stadium:
          type: string
          example: Camp nou
        superligaName:
          type: string
          example: Ligat Haal
        seasonName:
          type: string
          example: 2020/2021
        stageName:
          type: string
          example: Final
        refereeName:
        type: string
        example: Dvir Rainisch
        homeTeamID:
          type: integer
          example: 12345
        awayTeamID:
          type: integer
          example: 12345
    Match:
      type: object
      properties:
        previewMatch:
          $ref: '#/components/schemas/PreviewMatch'
        homeGoals:
          type: integer
          example: 3
        awayGoals:
          type: integer
          example: 2
        eventsCalendar:
          type: array
          items:
            $ref: '#/components/schemas/EventDetails'